@@ -1,41 +1,42 @@
 {
-    "title": "Snow Goat",
-    "name": "snow-goat",
-    "version": "0.0.1",
-    "description": "Home of ServiceNow's Greatest of all Time",
-    "keywords": [
-        "snow",
-        "goat",
-        "servicenow"
-    ],
-    "homepage": "https://snowgo.at",
-    "license": "MIT",
-    "author": "Tommy Ryan",
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/tallboy/snowgo.at"
-    },
-    "scripts": {
-        "start": "parcel ./index.ejs --open",
-        "clean": "del dist",
-        "prebuild": "npm run clean",
-        "build": "parcel build ./index.ejs",
-        "lint": "eslint .  --ext .js",
-        "lint-fix": "eslint --fix .  --ext .js"
-    },
-    "browserslist": "defaults",
-    "dependencies": {
-        "bootstrap": "5.1.3",
-        "font-awesome": "4.7.0",
-        "jquery": "3.7.1",
-        "jquery.easing": "^1.4.1"
-    },
-    "devDependencies": {
-        "@parcel/transformer-sass": "^2.12.0",
-        "del-cli": "^5.1.0",
-        "eslint": "8.53.0",
-        "eslint-plugin-import": "^2.25.2",
-        "parcel": "^2.12.0",
-        "parcel-transformer-ejs": "^1.0.1"
-    }
+  "title": "Snow Goat",
+  "name": "snow-goat",
+  "version": "0.0.1",
+  "description": "Home of ServiceNow's Greatest of all Time",
+  "type": "module",
+  "keywords": [
+    "snow",
+    "goat",
+    "servicenow"
+  ],
+  "homepage": "https://snowgo.at",
+  "license": "MIT",
+  "author": "Tommy Ryan",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/tallboy/snowgo.at"
+  },
+  "scripts": {
+    "start": "parcel ./index.ejs --open",
+    "clean": "del dist",
+    "prebuild": "npm run clean",
+    "build": "parcel build ./index.ejs",
+    "lint": "eslint .",
+    "lint-fix": "eslint --fix ."
+  },
+  "browserslist": "defaults",
+  "dependencies": {
+    "bootstrap": "5.1.3",
+    "font-awesome": "4.7.0",
+    "jquery": "3.7.1",
+    "jquery.easing": "^1.4.1"
+  },
+  "devDependencies": {
+    "@parcel/transformer-sass": "^2.12.0",
+    "del-cli": "^5.1.0",
+    "eslint": "^9.9.0",
+    "parcel": "^2.12.0",
+    "parcel-transformer-ejs": "^1.0.1",
+    "prettier": "^3.3.3"
+  }
 }
